** telescope bespoke cli args, + ATA snr and driftrate
