+++ apps/blade-cli/include/blade-cli/telescopes/ata/base.hh
@@ -1,7 +1,12 @@
 #ifndef BLADE_CLI_TELESCOPES_ATA_HH
 #define BLADE_CLI_TELESCOPES_ATA_HH
 
+#include <chrono>
+#include <CLI/CLI.hpp>
+
 #include "blade-cli/types.hh"
+
+#include "blade-cli/telescopes/ata/config.hh"
 
 #if defined(BLADE_PIPELINE_ATA_MODE_B)
 #include "blade-cli/telescopes/ata/mode_b.hh"
@@ -16,6 +21,7 @@
 
 namespace Blade::CLI::Telescopes::ATA {
 
+const Result CollectUserInput(int argc, char **argv, Config& config);
 const Result Setup(const Config& config);
 
 }  // namespace Blade::CLI::Telecopes::ATA
