+++ apps/blade-cli/src/telescopes/generic.cc
@@ -15,4 +15,10 @@ const Result Setup(const Config& config) {
     return Result::ERROR;
 }
 
+const Result CollectUserInput(int argc, char **argv, Config& config) {
+    ::CLI::App app("BLADE (Breakthrough Listen Accelerated DSP Engine) - Command Line Tool - Generic");
+
+    return Result::SUCCESS;
+}
+
 }  // namespace Blade::CLI::Telescopes::Generic
